Add files via upload
This project has been carried out as a part of a project assignment in the 'Data Science by Python' course. Using data available in "Coronavirus Source Data - Our World in Data," this study focuses on the total cases and the death cases in a few countries. The analysis has been done on June 29, 2020.